Optimize bundle splitting for using `next-intl/navigation` on the client side
