+++ packages/next-intl/src/react-client/useLocale.tsx
@@ -1,5 +1,6 @@
 import {useParams} from 'next/navigation';
-import {useLocale as useBaseLocale} from 'use-intl';
+// Workaround for some bundle splitting until we have ESM
+import useBaseLocale from 'use-intl/_useLocale';
 import {LOCALE_SEGMENT_NAME} from '../shared/constants';
 
 export default function useLocale(): string {
