+++ packages/next-intl/src/shared/NextIntlClientProvider.tsx
@@ -2,7 +2,8 @@
 
 import {useRouter} from 'next/router';
 import React, {ComponentProps} from 'react';
-import {IntlProvider} from 'use-intl';
+// Workaround for some bundle splitting until we have ESM
+import IntlProvider from 'use-intl/_IntlProvider';
 
 type Props = Omit<ComponentProps<typeof IntlProvider>, 'locale'> & {
   locale?: string;
