+++ packages/use-intl/_IntlProvider.d.ts
@@ -1,0 +1,1 @@
+export {default} from '../dist/_IntlProvider';
