+++ packages/use-intl/_useLocale.d.ts
@@ -1,0 +1,1 @@
+export {default} from '../dist/_useLocale';
