+++ packages/use-intl/package.json
@@ -11,7 +11,7 @@
     "url": "https://github.com/amannn/next-intl/tree/main/packages/use-intl"
   },
   "scripts": {
-    "build": "dts build --entry src/index.tsx && dts build --entry src/core.tsx --entry src/react.tsx --noClean",
+    "build": "dts build --entry src/index.tsx && dts build --entry src/core.tsx --entry src/react.tsx --entry src/_useLocale.tsx --entry src/_IntlProvider.tsx --noClean",
     "test": "TZ=Europe/Berlin vitest",
     "lint": "eslint src test && tsc --noEmit",
     "prepublishOnly": "CI=true turbo test && turbo lint && turbo build",
@@ -23,12 +23,16 @@
   "exports": {
     ".": "./dist/index.js",
     "./core": "./dist/core.js",
-    "./react": "./dist/react.js"
+    "./react": "./dist/react.js",
+    "./_useLocale": "./dist/_useLocale.js",
+    "./_IntlProvider": "./dist/_IntlProvider.js"
   },
   "files": [
     "dist",
     "core.d.ts",
-    "react.d.ts"
+    "react.d.ts",
+    "_useLocale.d.ts",
+    "_IntlProvider.d.ts"
   ],
   "keywords": [
     "react",
