+++ packages/use-intl/src/_IntlProvider.tsx
@@ -1,0 +1,1 @@
+export {default} from './react/IntlProvider';
