+++ packages/use-intl/src/_useLocale.tsx
@@ -1,0 +1,1 @@
+export {default} from './react/useLocale';
